2 21 3 (#484)
fix: more deprecation warnings on image uploads
fix: broken avatar upload when using the french language

+++ endpoints/payments/add.php
@@ -115,13 +115,13 @@ function resizeAndUploadLogo($uploadedFile, $uploadDir, $name)
             $newHeight = $height;
 
             if ($width > $targetWidth) {
-                $newWidth = $targetWidth;
-                $newHeight = ($targetWidth / $width) * $height;
+                $newWidth = (int)$targetWidth;
+                $newHeight = (int)(($targetWidth / $width) * $height);
             }
 
             if ($newHeight > $targetHeight) {
-                $newWidth = ($targetHeight / $newHeight) * $newWidth;
-                $newHeight = $targetHeight;
+                $newWidth = (int)(($targetHeight / $newHeight) * $newWidth);
+                $newHeight = (int)$targetHeight;
             }
 
             $resizedImage = imagecreatetruecolor($newWidth, $newHeight);
